+++ cmake/base_config.cmake
@@ -49,7 +49,7 @@ endif()
 set(NAMESPACE qb)
 
 # Set Compiler Support
-set(CMAKE_CXX_STANDARD 20 CACHE STRING "Adopted C++ standard.")
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)